extract stats calculation to codeowners package
